+++ core/src/main/java/com/ensody/reactivestate/AutoRun.kt
@@ -3,13 +3,13 @@ package com.ensody.reactivestate
 import kotlinx.coroutines.CoroutineScope
 
 /** Observer callback used by [autoRun] and [AutoRunner]. */
-typealias AutoRunCallback<T> = Resolver.() -> T
+public typealias AutoRunCallback<T> = Resolver.() -> T
 
 /** onChange callback used by [autoRun] and [AutoRunner]. */
-typealias AutoRunOnChangeCallback<T> = (AutoRunner<T>) -> Unit
+public typealias AutoRunOnChangeCallback<T> = (AutoRunner<T>) -> Unit
 
 /**
- * Watches observables for changes. Often useful to keep things in sync (e.g. [Scoped] -> UI).
+ * Watches observables for changes. Often useful to keep things in sync (e.g. [CoroutineScopeOwner] -> UI).
  *
  * This is a convenience function that immediately starts the [AutoRunner.run] cycle for you.
  *
@@ -22,9 +22,9 @@ typealias AutoRunOnChangeCallback<T> = (AutoRunner<T>) -> Unit
  * manually call [run].
  * @param [observer] The callback which is used to track the observables.
  */
-fun CoroutineScope.autoRun(
+public fun CoroutineScope.autoRun(
     onChange: AutoRunOnChangeCallback<Unit>? = null,
-    observer: AutoRunCallback<Unit>
+    observer: AutoRunCallback<Unit>,
 ): AutoRunner<Unit> =
     AutoRunner(this, onChange, observer).apply {
         disposeOnCompletionOf(this@autoRun)
@@ -32,12 +32,12 @@ fun CoroutineScope.autoRun(
     }
 
 /**
- * Watches observables for changes. Often useful to keep things in sync (e.g. [Scoped] -> UI).
+ * Watches observables for changes. Often useful to keep things in sync (e.g. [CoroutineScopeOwner] -> UI).
  *
  * This is a convenience function that immediately starts the [AutoRunner.run] cycle for you.
  *
  * Returns the underlying [AutoRunner]. To stop watching, you should call [AutoRunner.dispose].
- * The [AutoRunner] is automatically disposed when the [Scoped.scope] completes.
+ * The [AutoRunner] is automatically disposed when the [CoroutineScopeOwner.scope] completes.
  *
  * See [AutoRunner] for more details.
  *
@@ -45,22 +45,22 @@ fun CoroutineScope.autoRun(
  * manually call [run].
  * @param [observer] The callback which is used to track the observables.
  */
-fun Scoped.autoRun(
+public fun CoroutineScopeOwner.autoRun(
     onChange: AutoRunOnChangeCallback<Unit>? = null,
-    observer: AutoRunCallback<Unit>
+    observer: AutoRunCallback<Unit>,
 ): AutoRunner<Unit> =
     scope.autoRun(onChange, observer)
 
 /** Just the minimum interface needed for [Resolver]. No generic types. */
-abstract class BaseAutoRunner : AttachedDisposables {
+public abstract class BaseAutoRunner : AttachedDisposables {
     internal abstract val resolver: Resolver
 
-    abstract val autoRunnerScope: CoroutineScope
-    abstract fun triggerChange()
+    public abstract val autoRunnerScope: CoroutineScope
+    public abstract fun triggerChange()
 }
 
 /**
- * Watches observables for changes. Often useful to keep things in sync (e.g. [Scoped] -> UI).
+ * Watches observables for changes. Often useful to keep things in sync (e.g. [CoroutineScopeOwner] -> UI).
  *
  * Given an [observer], this class will automatically register itself as a listener and keep track
  * of the observables which [observer] depends on.
@@ -75,14 +75,14 @@ abstract class BaseAutoRunner : AttachedDisposables {
  * manually call [run] at any point (e.g. asynchronously) to change the tracked observables.
  * @param [observer] The callback which is used to track the observables.
  */
-class AutoRunner<T>(
+public class AutoRunner<T>(
     override val autoRunnerScope: CoroutineScope,
     onChange: AutoRunOnChangeCallback<T>? = null,
-    private val observer: AutoRunCallback<T>
+    private val observer: AutoRunCallback<T>,
 ) : BaseAutoRunner() {
-    override val attachedDisposables = DisposableGroup()
-    val listener: AutoRunOnChangeCallback<T> = onChange ?: { run() }
-    override var resolver = Resolver(this)
+    override val attachedDisposables: DisposableGroup = DisposableGroup()
+    public val listener: AutoRunOnChangeCallback<T> = onChange ?: { run() }
+    override var resolver: Resolver = Resolver(this)
 
     /** Stops watching observables. */
     override fun dispose() {
@@ -91,7 +91,7 @@ class AutoRunner<T>(
     }
 
     /** Calls [observer] and tracks its dependencies. */
-    fun run(): T = observe(observer)
+    public fun run(): T = observe(observer)
 
     override fun triggerChange() {
         listener(this)
@@ -116,7 +116,7 @@ class AutoRunner<T>(
 }
 
 /** Tracks observables for [AutoRunner]. */
-class Resolver(val autoRunner: BaseAutoRunner) {
+public class Resolver(public val autoRunner: BaseAutoRunner) {
     private val observables = mutableMapOf<Any, AutoRunnerObservable>()
 
     /**
@@ -128,9 +128,9 @@ class Resolver(val autoRunner: BaseAutoRunner) {
      * @param [underlyingObservable] The raw, underlying observable (e.g. Android's `LiveData`).
      * @param [getObservable] Used to create an [AutoRunnerObservable] wrapper around [underlyingObservable].
      */
-    fun <S : Any, T : AutoRunnerObservable> track(
+    public fun <S : Any, T : AutoRunnerObservable> track(
         underlyingObservable: S,
-        getObservable: () -> T
+        getObservable: () -> T,
     ): S {
         if (underlyingObservable !in observables) {
             val existing = autoRunner.resolver.observables[underlyingObservable]
@@ -157,7 +157,7 @@ class Resolver(val autoRunner: BaseAutoRunner) {
  *
  * You can use this to wrap actual observables (e.g. Android's `LiveData`).
  */
-interface AutoRunnerObservable {
-    fun addObserver()
-    fun removeObserver()
+public interface AutoRunnerObservable {
+    public fun addObserver()
+    public fun removeObserver()
 }