+++ reactivestate/src/main/java/com/ensody/reactivestate/BindingAndroid.kt
@@ -15,7 +15,7 @@ import kotlinx.coroutines.flow.StateFlow
 // -------------------------------------------------------------------------------------------------
 
 /** Keeps [data] in sync with [view]`.text`. */
-fun LifecycleOwner.bind(data: MutableLiveData<String>, view: TextView): Disposable =
+public fun LifecycleOwner.bind(data: MutableLiveData<String>, view: TextView): Disposable =
     DisposableGroup().apply {
         add(onStartOnce {
             val watcher = view.addTextChangedListener {
@@ -30,7 +30,7 @@ fun LifecycleOwner.bind(data: MutableLiveData<String>, view: TextView): Disposab
     }
 
 /** Keeps [data] in sync with [view]`.text`. */
-fun LifecycleOwner.bind(data: MutableStateFlow<String>, view: TextView): Disposable =
+public fun LifecycleOwner.bind(data: MutableStateFlow<String>, view: TextView): Disposable =
     DisposableGroup().apply {
         add(onStartOnce {
             val watcher = view.addTextChangedListener {
@@ -45,7 +45,7 @@ fun LifecycleOwner.bind(data: MutableStateFlow<String>, view: TextView): Disposa
     }
 
 /** Keeps [view]`.text` in sync with the value returned by [observer] (via [LifecycleOwner.autoRun]). */
-fun LifecycleOwner.bind(view: TextView, observer: AutoRunCallback<String>): Disposable =
+public fun LifecycleOwner.bind(view: TextView, observer: AutoRunCallback<String>): Disposable =
     autoRun {
         val value = observer()
         if (view.text.toString() != value) {
@@ -54,22 +54,22 @@ fun LifecycleOwner.bind(view: TextView, observer: AutoRunCallback<String>): Disp
     }
 
 /** Keeps [view]`.text` in sync with [data]. */
-fun LifecycleOwner.bind(view: TextView, data: LiveData<String>): Disposable =
+public fun LifecycleOwner.bind(view: TextView, data: LiveData<String>): Disposable =
     bind(view) { get(data) ?: "" }
 
 /** Keeps [view]`.text` in sync with [data]. */
-fun LifecycleOwner.bind(view: TextView, data: StateFlow<String>): Disposable =
+public fun LifecycleOwner.bind(view: TextView, data: StateFlow<String>): Disposable =
     bind(view) { get(data) }
 
 /** Keeps [data] and [view]`.text` in sync with each other (bidirectionally). */
-fun LifecycleOwner.bindTwoWay(data: MutableLiveData<String>, view: TextView): Disposable =
+public fun LifecycleOwner.bindTwoWay(data: MutableLiveData<String>, view: TextView): Disposable =
     DisposableGroup().apply {
         add(bind(view, data))
         add(bind(data, view))
     }
 
 /** Keeps [data] and [view]`.text` in sync with each other (bidirectionally). */
-fun LifecycleOwner.bindTwoWay(data: MutableStateFlow<String>, view: TextView): Disposable =
+public fun LifecycleOwner.bindTwoWay(data: MutableStateFlow<String>, view: TextView): Disposable =
     DisposableGroup().apply {
         add(bind(view, data))
         add(bind(data, view))
@@ -80,7 +80,7 @@ fun LifecycleOwner.bindTwoWay(data: MutableStateFlow<String>, view: TextView): D
 // -------------------------------------------------------------------------------------------------
 
 /** Keeps [view]`.isChecked` in sync with the value returned by [observer] (via [LifecycleOwner.autoRun]). */
-fun LifecycleOwner.bind(view: Checkable, observer: AutoRunCallback<Boolean>): Disposable =
+public fun LifecycleOwner.bind(view: Checkable, observer: AutoRunCallback<Boolean>): Disposable =
     autoRun {
         val value = observer()
         if (view.isChecked != value) {
@@ -89,11 +89,11 @@ fun LifecycleOwner.bind(view: Checkable, observer: AutoRunCallback<Boolean>): Di
     }
 
 /** Keeps [view]`.isChecked` in sync with [data]. If `data.value` is null, [default] is used. */
-fun LifecycleOwner.bind(view: Checkable, data: LiveData<Boolean>, default: Boolean = false): Disposable =
+public fun LifecycleOwner.bind(view: Checkable, data: LiveData<Boolean>, default: Boolean = false): Disposable =
     bind(view) { get(data) ?: default }
 
 /** Keeps [view]`.isChecked` in sync with [data]. If `data.value` is null, [default] is used. */
-fun LifecycleOwner.bind(view: Checkable, data: StateFlow<Boolean>): Disposable =
+public fun LifecycleOwner.bind(view: Checkable, data: StateFlow<Boolean>): Disposable =
     bind(view) { get(data) }
 
 // -------------------------------------------------------------------------------------------------
@@ -101,7 +101,7 @@ fun LifecycleOwner.bind(view: Checkable, data: StateFlow<Boolean>): Disposable =
 // -------------------------------------------------------------------------------------------------
 
 /** Keeps [data] in sync with [view]`.isChecked`. */
-fun LifecycleOwner.bind(data: MutableLiveData<Boolean>, view: CompoundButton): Disposable =
+public fun LifecycleOwner.bind(data: MutableLiveData<Boolean>, view: CompoundButton): Disposable =
     DisposableGroup().apply {
         add(onStartOnce {
             view.setOnCheckedChangeListener { _, isChecked ->
@@ -115,7 +115,7 @@ fun LifecycleOwner.bind(data: MutableLiveData<Boolean>, view: CompoundButton): D
     }
 
 /** Keeps [data] in sync with [view]`.isChecked`. */
-fun LifecycleOwner.bind(data: MutableStateFlow<Boolean>, view: CompoundButton): Disposable =
+public fun LifecycleOwner.bind(data: MutableStateFlow<Boolean>, view: CompoundButton): Disposable =
     DisposableGroup().apply {
         add(onStartOnce {
             view.setOnCheckedChangeListener { _, isChecked ->
@@ -129,26 +129,26 @@ fun LifecycleOwner.bind(data: MutableStateFlow<Boolean>, view: CompoundButton): 
     }
 
 /** Keeps [view]`.isChecked` in sync with the value returned by [observer] (via [LifecycleOwner.autoRun]). */
-fun LifecycleOwner.bind(view: CompoundButton, observer: AutoRunCallback<Boolean>): Disposable =
+public fun LifecycleOwner.bind(view: CompoundButton, observer: AutoRunCallback<Boolean>): Disposable =
     bind(view as Checkable, observer)
 
 /** Keeps [view]`.isChecked` in sync with [data]. If `data.value` is null, [default] is used. */
-fun LifecycleOwner.bind(view: CompoundButton, data: LiveData<Boolean>, default: Boolean = false): Disposable =
+public fun LifecycleOwner.bind(view: CompoundButton, data: LiveData<Boolean>, default: Boolean = false): Disposable =
     bind(view as Checkable, data, default = default)
 
 /** Keeps [view]`.isChecked` in sync with [data] */
-fun LifecycleOwner.bind(view: CompoundButton, data: StateFlow<Boolean>): Disposable =
+public fun LifecycleOwner.bind(view: CompoundButton, data: StateFlow<Boolean>): Disposable =
     bind(view as Checkable, data)
 
 /** Keeps [data] and [view]`.isChecked` in sync with each other (bidirectionally). */
-fun LifecycleOwner.bindTwoWay(data: MutableLiveData<Boolean>, view: CompoundButton): Disposable =
+public fun LifecycleOwner.bindTwoWay(data: MutableLiveData<Boolean>, view: CompoundButton): Disposable =
     DisposableGroup().apply {
         add(bind(view, data))
         add(bind(data, view))
     }
 
 /** Keeps [data] and [view]`.isChecked` in sync with each other (bidirectionally). */
-fun LifecycleOwner.bindTwoWay(data: MutableStateFlow<Boolean>, view: CompoundButton): Disposable =
+public fun LifecycleOwner.bindTwoWay(data: MutableStateFlow<Boolean>, view: CompoundButton): Disposable =
     DisposableGroup().apply {
         add(bind(view, data))
         add(bind(data, view))