+++ reactivestate/src/main/java/com/ensody/reactivestate/ViewModelAndroid.kt
@@ -11,7 +11,7 @@ import androidx.lifecycle.*
  * The [provider] should instantiate the `ViewModel` directly.
  */
 @Suppress("UNCHECKED_CAST")
-inline fun <reified T : ViewModel> Fragment.viewModel(crossinline provider: () -> T) =
+public inline fun <reified T : ViewModel> Fragment.buildViewModel(crossinline provider: () -> T): Lazy<T> =
     viewModels<T> {
         object : ViewModelProvider.Factory {
             override fun <T : ViewModel?> create(modelClass: Class<T>): T = provider() as T
@@ -24,7 +24,7 @@ inline fun <reified T : ViewModel> Fragment.viewModel(crossinline provider: () -
  * The [provider] should instantiate the `ViewModel` directly.
  */
 @Suppress("UNCHECKED_CAST")
-inline fun <reified T : ViewModel> Fragment.activityViewModel(crossinline provider: () -> T) =
+public inline fun <reified T : ViewModel> Fragment.activityViewModel(crossinline provider: () -> T): Lazy<T> =
     activityViewModels<T> {
         object : ViewModelProvider.Factory {
             override fun <T : ViewModel?> create(modelClass: Class<T>): T = provider() as T
@@ -37,13 +37,15 @@ inline fun <reified T : ViewModel> Fragment.activityViewModel(crossinline provid
  * The [provider] should instantiate the `ViewModel` directly.
  */
 @Suppress("UNCHECKED_CAST")
-inline fun <reified T : ViewModel> Fragment.stateViewModel(crossinline provider: (handle: SavedStateHandle) -> T) =
+public inline fun <reified T : ViewModel> Fragment.stateViewModel(
+    crossinline provider: (handle: SavedStateHandle) -> T,
+): Lazy<T> =
     viewModels<T> {
         object : AbstractSavedStateViewModelFactory(this, null) {
             override fun <T : ViewModel?> create(
                 key: String,
                 modelClass: Class<T>,
-                handle: SavedStateHandle
+                handle: SavedStateHandle,
             ): T = provider(handle) as T
         }
     }
@@ -54,13 +56,15 @@ inline fun <reified T : ViewModel> Fragment.stateViewModel(crossinline provider:
  * The [provider] should instantiate the `ViewModel` directly.
  */
 @Suppress("UNCHECKED_CAST")
-inline fun <reified T : ViewModel> Fragment.activityStateViewModel(crossinline provider: (handle: SavedStateHandle) -> T) =
+public inline fun <reified T : ViewModel> Fragment.activityStateViewModel(
+    crossinline provider: (handle: SavedStateHandle) -> T,
+): Lazy<T> =
     activityViewModels<T> {
         object : AbstractSavedStateViewModelFactory(requireActivity(), null) {
             override fun <T : ViewModel?> create(
                 key: String,
                 modelClass: Class<T>,
-                handle: SavedStateHandle
+                handle: SavedStateHandle,
             ): T = provider(handle) as T
         }
     }