+++ reactivestate/src/test/java/com/ensody/reactivestate/BindingTest.kt
@@ -12,7 +12,7 @@ import org.junit.runner.RunWith
 
 @ExperimentalCoroutinesApi
 @RunWith(AndroidJUnit4::class)
-class BindingTest {
+internal class BindingTest {
     @Test
     fun bindingOnFragment() {
         val scenario = launchFragmentInContainer<TestFragment>()
@@ -23,31 +23,31 @@ class BindingTest {
         lateinit var fragment: TestFragment
         scenario.onFragment { fragment = it }
 
-        fragment.apply {
-            bindTwoWay(state.name, textView)
-            state.count.value = 1
-            state.name.value = "test"
-            assertThat(textView.text.toString()).isEqualTo(state.name.value)
+        fragment.run {
+            bindTwoWay(viewModel.name, textView)
+            viewModel.count.value = 1
+            viewModel.name.value = "test"
+            assertThat(textView.text.toString()).isEqualTo(viewModel.name.value)
             textView.text = "hello"
-            assertThat(textView.text.toString()).isEqualTo(state.name.value)
+            assertThat("" + textView.text).isEqualTo(viewModel.name.value)
         }
 
         // Bindings should auto-dispose themselves
         scenario.moveToState(Lifecycle.State.CREATED)
-        fragment.apply {
-            state.name.value = "test"
-            assertThat(textView.text.toString()).isNotEqualTo(state.name.value)
+        fragment.run {
+            viewModel.name.value = "test"
+            assertThat(textView.text.toString()).isNotEqualTo(viewModel.name.value)
         }
 
         // Re-creating the fragment should keep view model state
         scenario.recreate()
         scenario.onFragment { fragment = it }
         scenario.moveToState(Lifecycle.State.RESUMED)
-        fragment.apply {
-            bindTwoWay(state.name, textView)
-            assertThat(state.name.value).isEqualTo("test")
-            assertThat(textView.text.toString()).isEqualTo(state.name.value)
-            assertThat(state.count.value).isEqualTo(1)
+        fragment.run {
+            bindTwoWay(viewModel.name, textView)
+            assertThat(viewModel.name.value).isEqualTo("test")
+            assertThat(textView.text.toString()).isEqualTo(viewModel.name.value)
+            assertThat(viewModel.count.value).isEqualTo(1)
         }
     }
 }